requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
bcrypt
python-jose[cryptography]
structlog
alembic

